Emulates arith.fptoui for wide integers. vector<*xi2N> is emulated as vector<*x2xiN>. scalars are also emulated with vector<2xiN>.
This pass, as of now only covers the basic case of arith.fptoui fN to
i2N. It includes a arith.bitcast that is not emulated, if the fp width
is greater than N. For supporting the non-basic case, one needs to
emulate the bitcast operation as well.

Also, the edge cases for -inf, +inf, 0, -0 and NaN are to be created as
well.