cherry-pick: [mlir][linalg] Fix partial fuse by collapse (#9)
Similar to `FoldWithProducerReshapeOpByCollapsing`,
`FoldReshapeWithGenericOpByCollapsing` needs to be able to handle
partial fusion of a reshape by collapsing. This means that the source of
the generated `expand_shape` op (aka the collapsed linalg op) might not
match the type of the original `collapse_shape` op. This change instead
replaces the original linalg op with the new `expand_shape` op which is
guaranteed to be the same type.

This is a cherry-pick of commit 4652b69b0a512ff248b08cfa7ba9547860d1cc10.