fix(mlir,tosa): restore unrealized casts when lowering `tosa.rescale` to linalg
Along with the changes to rescale op attributes, commit 7208649 dropped the builtin casts
between signed and signless types. However, explicitly unsigned types are still legal input
and output values from the TOSA IR perspective, and TF-to-TOSA's `ConvertUint8ToInt8` pass
generates just that.

The change adds back the casts when the unsigned<->signless semantics are explicit in the
underlying tensor types. This prevents the conversion routine from trying to generate illegal
`arith` casts that are contrained to signless types. Whether the `arith` casts themselves are
signed or unsigned should still depend on the rescale's `*_unsigned` attribute values.