feat(linalg): add a way to pass controlFn to `foldIntoPackUnpackPatterns` (#22)
This is step 1/3 of fixing the large vector sizes found compilation failure (roo-389). This PR adds a mechanism, so that downstream consumers can pass in control functions for the application of these patterns. This change shouldn't affect any consumers of this method that do not specify a controlFn.

In IREE, we (will) use it to control preventing folding patterns that would inhibit fusion. See IREE issue #20896 for more details.